MEMORIES OF MY FAMILY MEETINGS still are a source of strength to me.  I remember we'd all get into the car -- I forget what kind it was -- and drive and drive.
I'm not sure where we'd go, but I think there were some bees there. The
smell of something was strong in the air as we played whatever sport we
played.  I remember a bigger, older guy whom we called "Dad."  We'd eat
some stuff or not and then I think we went home.

I guess some things never leave you.
		-- Jack Handley, The New Mexican, 1988.